backend/v1/serializers/products.py
from rest_framework import serializers
from v1.models.products import Products

class ProductsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'