backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-compat==1.0.15
django-cors-headers==4.3.1
django-filter==24.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-standardized-errors==0.14.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
